requirements.txt
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.3
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djongo==1.3.6
sqlparse==0.2.4
openapi-codec==1.3.2
psycopg2-binary==2.8.6
pymongo==3.11.4
rest-framework-dyn-serializer==1.3.1
python-memcached==1.59.0
jaeger-django==1.0.2
django-jaeger-middleware==1.0.0
django-opentracing==1.1.0
opentracing==2.4.0
